REFACTOR(approval-request): moved to /accounts (#282)
### TL;DR
Restructured the approval request feature by moving components into a dedicated accounts subdirectory and updated navigation paths.

### What changed?
- Moved approval request components into a new `/accounts` subdirectory
- Updated import paths across components to reflect new directory structure
- Added new `ApproveRequestButton` component
- Changed navigation item text from "Approval Requests" to "Accounts Requests"
- Updated navigation href from `/admin/approval-request` to `/admin/approval-request/accounts`

### How to test?
1. Navigate to the admin dashboard
2. Verify the navigation menu shows "Accounts Requests"
3. Click on "Accounts Requests" and confirm it routes to the correct page
4. Verify all approval request functionality continues to work as expected

### Why make this change?
This restructuring prepares the codebase for handling different types of approval requests beyond just accounts, improving scalability and maintainability of the approval request feature.
