+++ src/app/(dashboard)/admin/approval-request/accounts/approve-request-button.tsx
@@ -1,0 +1,12 @@
+import { Button } from '@/components/ui/button'
+import { FC, ReactNode } from 'react'
+
+interface ApproveRequestButtonProps {
+  children: ReactNode
+}
+
+const ApproveRequestButton: FC<ApproveRequestButtonProps> = ({ children }) => {
+  return <div>{children}</div>
+}
+
+export default ApproveRequestButton
