+++ src/app/(dashboard)/admin/approval-request/approve-request-button.tsx
@@ -1,0 +1,11 @@
+import { FC, ReactNode } from 'react'
+
+interface ApproveRequestButtonProps {
+  children: ReactNode
+}
+
+const ApproveRequestButton: FC<ApproveRequestButtonProps> = ({ children }) => {
+  return <div>{children}</div>
+}
+
+export default ApproveRequestButton
